pyproject.toml
[tool.poetry]
name = "candycard"
version = "0.1.0"
description = ""
authors = ["misano <me@MirSobhan.ir>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.13"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
cachetools = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
